prepare for change city::states

+++ source/good/helper.cpp
@@ -159,7 +159,7 @@ float Helper::importPrice(PlayerCityPtr city, Product gtype, unsigned int qty)
 {
   int price = city->tradeOptions().sellPrice(gtype);
   Unit units = Unit::fromQty(qty);
-  return price * units.ivalue();
+  return (float)(price * units.ivalue());
 }
 
 Product Helper::random()
